Fix plugin order
In the `get_plugins` method of the placeholder class, the CMS orders
the plugins by `path`. When using the placeholder template tag, one of
the last steps before rendering is a sort method after `position`. We
didn't realize that until now and needed to fix it with this commit.

Related issue:
https://github.com/divio/django-cms/pull/5153